Support python2
Running obs-to-maven on Debian wouldn't fly as of today due to missing
python3 dependencies. For instance osc is python2 and requires
python-m2crypto which has been packaged for python 3 on Debian 11+.

On the other hand rpmfile has some troubles to run on python2 due to
LZMA. The pyliblzma module packaged on debian doesn't have a stream
interface like the python 3.3 lzma module used in rpmfile. For this
reason, rpmfile is dropped in favor or python's rpm binding combined
with rpm2cpio | cpio.