Use streams to decompress and parse the primary XML
Primary XML files can grow really huge depending on the number of RPMs
of the repo. Uncompressing the whole file to a string and then parse the
string costs a lot of memory.

For the comparison, the current Leap 15.3 primary.xml weights 25MB
compressed. Using the string to parse it takes 286MB of RAM, while
switching to parse streams the consumed memory goes down to 26MB.