Download primary to disk first
We see a lot of ConnectionResetErrors in our JenkinsCI. I hope that
downloading the whole compressed file first will help with that, based
on the assumption that the server resets the connection when the current
stream-based parsing takes too long.

We will still decompress the file as we parse without reading it into
memory to keep the memory footprint low.